backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart
python-dotenv
pydantic>=2.6.0
pydantic-settings
requests
httpx
Pillow>=10.2.0
numpy
# rembg>=2.0.50  # Disabled due to Python 3.14 compat
# onnxruntime>=1.17.0 # Disabled
diffusers>=0.26.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless
playwright
# google-generativeai
tenacity
structlog
openai
